requirements.txt
torch
torchvision
numpy
matplotlib
opencv-python
scikit-learn
tqdm
flask
flask-cors
pillow
